Use types consistent with QValueList::size(), so kMin() works out right.
svn path=/trunk/koffice/; revision=309997

+++ lib/store/koxmlwriter.cc
@@ -22,7 +22,7 @@
 #include <kdebug.h>
 #include <float.h>
 
-static const unsigned int s_indentBufferLength = 100;
+static const size_t s_indentBufferLength = 100;
 
 KoXmlWriter::KoXmlWriter( QIODevice* dev )
     : m_dev( dev )
